# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Leitstand API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/missions/": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "List Missions",
        "description": "List missions newest first, each with its latest run.\n\nA mission is a definition that can be run any number of times; ``latest_run`` carries the\nstatus, robot and timings of its most recent run and is null when it has never run. Its\nstatus is one of PENDING, DISPATCHED, RUNNING, PAUSED, SUCCEEDED, FAILED, CANCELLED or\nREJECTED. Optionally filter by robot, by name, or both. This returns definitions, not live\nper-stage progress: use get_mission_state for that.",
        "operationId": "list_missions",
        "parameters": [
          {
            "name": "robot",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to missions whose default robot is this robot id.",
              "title": "Robot"
            },
            "description": "Filter to missions whose default robot is this robot id."
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to missions with exactly this name, matched case-insensitively.",
              "title": "Name"
            },
            "description": "Filter to missions with exactly this name, matched case-insensitively."
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return archived missions.",
              "default": false,
              "title": "Include Archived"
            },
            "description": "Also return archived missions."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MissionView"
                  },
                  "title": "Response List Missions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Create Mission",
        "description": "Create a mission from a name and an ordered list of navigation stages.\n\nEvery waypoint must be a coordinate the operator stated. Do not compute one: not from a field\nor site boundary, not from a robot's current position, and not by converting a distance in\nmetres into degrees. If you were given an area, a row spacing or a bearing rather than\ncoordinates, do not call this: use plan_coverage_mission for a field that should be covered, and\notherwise say you cannot work the coordinates out and ask for them.\n\n``stages`` is a list of stage objects, not text containing a list.\n\nThe backend assigns the mission and stage ids. The mission is a definition and is not\ndispatched; dispatch_mission runs it, as many times as wanted.",
        "operationId": "create_mission",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Get Mission",
        "description": "Return one mission's definition, default robot and latest run.\n\nIdentify the mission by mission_id from list_missions, which is also how a mission named by\nthe operator is resolved. Its run history is list_mission_runs.",
        "operationId": "get_mission",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "missions"
        ],
        "summary": "Update Mission",
        "description": "Change a mission's name, description, or stages.\n\nAllowed at any time, even while a run is active: a run carries its own copy of the plan, so\nediting the definition never changes what a run did or is doing. Identify the mission by\nmission_id from list_missions. Supplying stages replaces the existing ones; a stage that\ncarries its stage_id keeps its identity across the edit, one without gets a new id, and one\nleft out is removed. A planned coverage stage is carried by its stage_id alone; its path\ncannot be written here, only re-planned.",
        "operationId": "update_mission",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "missions"
        ],
        "summary": "Delete Mission",
        "description": "Delete a mission that never ran, or archive one that did.\n\nA mission with runs is archived rather than deleted: it leaves the list, its runs stay\nreadable, and restore_mission brings it back. A mission with a run in progress is refused.\nIdentify the mission by mission_id from list_missions.",
        "operationId": "delete_mission",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/state": {
      "get": {
        "tags": [
          "missions"
        ],
        "summary": "Get Mission State",
        "description": "Return live per-stage progress of the mission's current run, errors attributed per stage.\n\nThe current run is the most recently started one still active, else the latest run of any\noutcome. Each stage carries a status and a progress fraction. Use this to answer how far\nalong a mission is, or why it failed. A mission that has never run has no state (404).",
        "operationId": "get_mission_state",
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStateView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/coverage": {
      "post": {
        "tags": [
          "missions"
        ],
        "summary": "Plan Coverage Mission",
        "description": "Create a mission that covers a whole field, planned from the field's own boundary.\n\nUse this whenever the operator asks to cover, survey, mow or treat a field rather than to\ndrive to stated points. Give the field_id, the robot the plan is for, and the working width\nin metres; supply no coordinates, because the path is computed from the stored boundary, and\nno turning radius, because the robot declares its own.\n\nWith ``replan`` set, the named mission's plan is rewritten in place instead and its id is\nreturned. The mission is not dispatched. Its coverage metrics come back with it, so the\noperator can judge the plan before dispatching it.",
        "operationId": "plan_coverage_mission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCoverageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "stages": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/NavigationStageInput"
                },
                {
                  "$ref": "#/components/schemas/CoverageStageRef"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "coverage": "#/components/schemas/CoverageStageRef",
                  "navigation": "#/components/schemas/NavigationStageInput"
                }
              }
            },
            "type": "array",
            "minItems": 1,
            "title": "Stages"
          }
        },
        "type": "object",
        "required": [
          "name",
          "stages"
        ],
        "title": "MissionCreate"
      },
...
      "RunStateView": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "mission_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mission Id"
          },
          "stage_states": {
            "items": {
              "$ref": "#/components/schemas/StageStateView"
            },
            "type": "array",
            "title": "Stage States"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "mission_id",
          "stage_states"
        ],
        "title": "RunStateView",
        "description": "A run's per-stage state as one read shape for both REST and the live WS frame.\n\nCarries both ids: the WS topic is keyed by mission, so a subscriber showing one run must\nfilter frames by ``run_id``, and a list of missions indexes them by ``mission_id``."
      },
...
    }
  }
}